requirements.txt
numpy==1.26.4
tensorflow==2.16.1
pandas==2.2.1
transformers==4.39.2
scipy==1.12.0
scikit-scikit-learn==1.4.1.post1
spacy==3.7.4
pillow==10.2.0
Pattern==3.6
munch==4.0.0
ipywidgets==8.1.2
iso-639==0.4.5
python-dotenv==1.0.1
sentencepiece==0.2.0
tf_keras==2.16.0
torch==2.2.2
